+++ dev/ci/user-overlays/14846-herbelin-master+inlining-inductive-names-in-inductive-blocks.sh
@@ -1,0 +1,5 @@
+overlay coq_dpdgraph https://github.com/herbelin/coq-dpdgraph.git coq-master+adapt-coq-pr14846-ind-as-ind-rather-than-rel-in-constructor-types 14846 master+inlining-inductive-names-in-inductive-blocks
+overlay elpi https://github.com/herbelin/coq-elpi coq-master+adapt-coq-pr14846-ind-as-ind-rather-than-rel-in-constructor-types 14846 master+inlining-inductive-names-in-inductive-blocks
+overlay equations https://github.com/herbelin/Coq-Equations master+adapt-coq-pr14846-ind-as-ind-rather-than-rel-in-constructor-types 14846 master+inlining-inductive-names-in-inductive-blocks
+overlay paramcoq https://github.com/herbelin/paramcoq master+adapt-coq-pr14846-ind-as-ind-rather-than-rel-in-constructor-types 14846 master+inlining-inductive-names-in-inductive-blocks
+overlay metacoq https://github.com/herbelin/template-coq master+adapt-coq-pr14846-ind-as-ind-rather-than-rel-in-constructor-types 14846 master+inlining-inductive-names-in-inductive-blocks